requirements.txt
pillow>=9.0.0
pdf2image>=1.16.0
numpy>=1.19.0
matplotlib>=3.5.0
pytesseract>=0.3.9
paddlepaddle>=2.4.0
paddleocr>=2.6.0
tk
pyinstaller>=5.7.0